worker/requirements.txt
celery
redis
fleep
srt
pympi-ling
yt-dlp
pydub
ffmpeg-normalize
SciPy
webvtt-py
anaouder